requirements.txt
# Data Handling
pandas
numpy

# Financial Indicators
pandas-ta

# Machine Learning & AI
scikit-learn
joblib

# Dashboard & Frontend
streamlit
plotly

# Database
supabase

# Utilities
yfinance
python-dotenv